+++ src/component/LoginWrapper/LoginWrapper.tsx
@@ -1,19 +1,23 @@
 import React, { PropsWithChildren } from "react";
-import { useUser } from "../../provider";
-import { Button, ButtonProps, Optional } from "@encode42/mantine-extras";
+import { Button, ButtonProps } from "@encode42/mantine-extras";
 import { Link, LinkProps } from "../Link";
-import { AvatarMenu, AvatarMenuProps } from "./AvatarMenu";
+import { useUser } from "../../provider";
 
 /**
- * Options for the {@link AvatarMenuWrapper} component.
+ * Options for the {@link LoginWrapper} component.
  */
-export interface AvatarMenuWrapperProps extends AvatarMenuProps, Optional<PropsWithChildren, "children"> {
+export interface LoginWrapperProps extends PropsWithChildren {
     /**
      * Where the logout button links to.
      *
      * Defaults to {@code /login}.
      */
     "to"?: LinkProps["to"],
+
+    /**
+     * Label of the login button.
+     */
+    "label"?: string,
 
     /**
      * Options for the {@link Link} component.
@@ -29,16 +33,16 @@ export interface AvatarMenuWrapperProps extends AvatarMenuProps, Optional<PropsW
 /**
  * Wrapper for the {@link AvatarMenu} component that displays a login button when logged out.
  */
-export function AvatarMenuWrapper({ to = "/login", linkProps, buttonProps, children = "Login", ...other }: AvatarMenuWrapperProps) {
+export function LoginWrapper({ to = "/login", linkProps, buttonProps, label = "Login", children }: LoginWrapperProps) {
     const { user } = useUser();
 
     return (
         user ? (
-            <AvatarMenu {...other} />
+            {children}
         ) : (
             <Link to={to} {...linkProps}>
                 <Button {...buttonProps}>
-                    {children}
+                    {label}
                 </Button>
             </Link>
         )