+++ src/component/index.ts
@@ -1,5 +1,5 @@
 export * from "./Anchor";
-export * from "./AvatarMenu";
 export * from "./ErrorPage";
 export * from "./Link";
 export * from "./Login";
+export * from "./LoginWrapper";